fix: correct blockIndex ABI type from uint256 to uint8 in fillChunkBlock(s)
The Go IStorageFillChunkBlockArgs struct defines BlockIndex as uint8.
Using uint256 in the ABI produces an incorrect function selector,
causing transaction failures on-chain.

Fix both fillChunkBlock and fillChunkBlocks ABI entries in storage.py.

+++ private/ipc/contracts/storage.py
@@ -1530,9 +1530,9 @@ class StorageContract:
                                 "type": "uint256"
                             },
                             {
-                                "internalType": "uint256",
+                                "internalType": "uint8",
                                 "name": "blockIndex",
-                                "type": "uint256"
+                                "type": "uint8"
                             },
                             {
                                 "internalType": "string",
@@ -1590,9 +1590,9 @@ class StorageContract:
                                 "type": "uint256"
                             },
                             {
-                                "internalType": "uint256",
+                                "internalType": "uint8",
                                 "name": "blockIndex",
-                                "type": "uint256"
+                                "type": "uint8"
                             },
                             {
                                 "internalType": "string",
